+++ pocketbase_helpers/lib/pocketbase_helpers.dart
@@ -7,3 +7,4 @@ export 'src/helper_utils.dart';
 export 'src/pocketbase_connection.dart';
 export 'src/auth_helper.dart';
 export 'src/file_helpers.dart';
+export 'src/realtime_helper.dart';