+++ pocketbase_helpers/test/auth_helper_test.dart
@@ -42,7 +42,7 @@ void main() {
         'password',
       );
 
-      expect(result, equals(AuthStatus.ok));
+      expect(result.status, equals(AuthStatus.ok));
       expect(result.record?.id, equals(expected.id));
     });